Separated HeadlessTests out from VisualTests; added test case for HoldNote key transfer
