+++ S2VX.Game.Tests/UnitTests/CommandTests.cs
@@ -5,7 +5,7 @@ using osuTK.Graphics;
 using S2VX.Game.Story.Command;
 using System;
 
-namespace S2VX.Game.Tests.HeadlessTests {
+namespace S2VX.Game.Tests.UnitTests {
     [TestFixture]
     public class CommandTests {
