+++ S2VX.Game.Tests/UnitTests/InterpolationTests.cs
@@ -3,7 +3,7 @@ using osu.Framework.Graphics;
 using osuTK;
 using osuTK.Graphics;
 
-namespace S2VX.Game.Tests.HeadlessTests {
+namespace S2VX.Game.Tests.UnitTests {
     [TestFixture]
     public class InterpolationTests {
