+++ S2VX.Game.Tests/VisualTests/EditorNoteTests.cs
@@ -32,70 +32,6 @@ namespace S2VX.Game.Tests.VisualTests {
         }
 
         [Test]
-        public void EditorNoteAlpha_BeforeFadeInTime_IsZero() {
-            AddStep("Seek before FadeInTime", () => StoryClock.Seek(NoteAppearTime));
-            AddAssert("Note is not visible", () => NoteToTest.Alpha == 0);
-        }
-
-        [Test]
-        public void EditorNoteAlpha_AfterFadeInBeforeShowTime_IsBetweenZeroAndOne() {
-            AddStep("Seek between FadeInTime and ShowTime", () => StoryClock.Seek(NoteAppearTime + Story.Notes.FadeInTime / 2));
-            AddAssert("Note is partially visible", () => NoteToTest.Alpha is > 0 and < 1);
-        }
-
-        [Test]
-        public void EditorNoteAlpha_AfterShowTimeBeforeHitTime_IsOne() {
-            AddStep("Seek between ShowTime and HitTime", () => StoryClock.Seek(NoteAppearTime + Story.Notes.FadeInTime + Story.Notes.ShowTime / 2));
-            AddAssert("Note is fully visible", () => NoteToTest.Alpha == 1);
-        }
-
-        [Test]
-        public void EditorNoteAlpha_AfterHitTimeBeforeFadeOutTime_IsBetweenZeroAndOne() {
-            AddStep("Seek between HitTime and FadeOutTime", () =>
-                StoryClock.Seek(NoteAppearTime + Story.Notes.FadeInTime + Story.Notes.ShowTime + Story.Notes.FadeOutTime / 2));
-            AddAssert("Note is partially visible", () => NoteToTest.Alpha is > 0 and < 1);
-        }
-
-        [Test]
-        public void EditorNoteAlpha_AfterFadeOutTime_IsZero() {
-            AddStep("Seek after FadeOutTime", () =>
-                StoryClock.Seek(NoteAppearTime + Story.Notes.FadeInTime + Story.Notes.ShowTime + Story.Notes.FadeOutTime));
-            AddAssert("Note is not visible", () => NoteToTest.Alpha == 0);
-        }
-
-        [Test]
-        public void EditorNoteApproachAlpha_BeforeFadeInTime_IsZero() {
-            AddStep("Seek before FadeInTime", () => StoryClock.Seek(NoteAppearTime));
-            AddAssert("Note approach is not visible", () => NoteToTest.Approach.Alpha == 0);
-        }
-
-        [Test]
-        public void EditorNoteApproachAlpha_AfterFadeInBeforeShowTime_IsBetweenZeroAndOne() {
-            AddStep("Seek between FadeInTime and ShowTime", () => StoryClock.Seek(NoteAppearTime + Story.Notes.FadeInTime / 2));
-            AddAssert("Note approach is partially visible", () => NoteToTest.Approach.Alpha is > 0 and < 1);
-        }
-
-        [Test]
-        public void EditorNoteApproachAlpha_AfterShowTimeBeforeHitTime_IsOne() {
-            AddStep("Seek between ShowTime and HitTime", () => StoryClock.Seek(NoteAppearTime + Story.Notes.FadeInTime + Story.Notes.ShowTime / 2));
-            AddAssert("Note approach is fully visible", () => NoteToTest.Approach.Alpha == 1);
-        }
-
-        [Test]
-        public void EditorNoteApproachAlpha_AfterHitTimeBeforeFadeOutTime_IsBetweenZeroAndOne() {
-            AddStep("Seek between HitTime and FadeOutTime", () =>
-                StoryClock.Seek(NoteAppearTime + Story.Notes.FadeInTime + Story.Notes.ShowTime + Story.Notes.FadeOutTime / 2));
-            AddAssert("Note approach is partially visible", () => NoteToTest.Approach.Alpha is > 0 and < 1);
-        }
-
-        [Test]
-        public void EditorNoteApproachAlpha_AfterFadeOutTime_IsZero() {
-            AddStep("Seek after FadeOutTime", () =>
-                StoryClock.Seek(NoteAppearTime + Story.Notes.FadeInTime + Story.Notes.ShowTime + Story.Notes.FadeOutTime));
-            AddAssert("Note approach is not visible", () => NoteToTest.Approach.Alpha == 0);
-        }
-
-        [Test]
         public void Hit_BeforeHitTime_DoesNotPlay() =>
             AddAssert("Does not play", () => NoteToTest.Hit.PlayCount == 0);
